Extract sub components from BaseColorsSection
