+++ src/components/layout/BaseColorsSection/index.tsx
@@ -1,30 +1,23 @@
 import { type ComponentProps, forwardRef, useState } from 'react';
 
+import ColorInputWithRandomBtn from './ColorInputWithRandomBtn';
+import ColorSuggestionsBox from './ColorSuggestionsBox';
 import {
   AccordionList,
   AccordionListItem,
   Collapsible,
-  IconButton,
   ListItem,
   ListSubheader,
   Radio,
 } from 'components/general';
-import { ColorInput, type ColorInputProps } from 'components/colors';
 import { useGlobalState } from 'hooks';
-import {
-  getDangerColor,
-  getNeutralColor,
-  isHexColorLight,
-  randomHexColor,
-} from 'utils';
-import { DoneIcon, RandomIcon } from 'assets/icons';
+import { getDangerColor, getNeutralColor } from 'utils';
 import type { GlobalState } from 'contexts/globalState';
 import {
   dangerColorSuggestionNames,
   dangerColorSuggestions,
   neutralColorSuggestionNames,
   neutralColorSuggestions,
-  type ColorSuggestions,
   type DangerColorSuggestion,
   type NeutralColorSuggestion,
 } from 'constants/colorSuggestions';
@@ -39,56 +32,6 @@ const ListItemRadio = ({
   <Radio checked={checked} disabled={disabled} className='me-4' asChild>
     <span />
   </Radio>
-);
-
-const ColorInputWithRandomBtn = ({
-  value,
-  onChange,
-  ...props
-}: ColorInputProps) => (
-  <div className='p-2'>
-    <ColorInput
-      {...props}
-      value={value}
-      onChange={onChange}
-      endAdornment={
-        <IconButton
-          title='Generate random color'
-          onClick={() => onChange?.(randomHexColor())}
-        >
-          <RandomIcon />
-        </IconButton>
-      }
-    />
-  </div>
-);
-
-const ColorSuggestionsBox = <T extends ColorSuggestions>({
-  baseColor,
-  colorSuggestions,
-  selectedSuggestion,
-  onSuggestionSelect,
-}: {
-  baseColor: string;
-  colorSuggestions: T;
-  selectedSuggestion?: keyof T;
-  onSuggestionSelect?: (suggestionName: keyof T) => void;
-}) => (
-  <div className='flex flex-row flex-wrap gap-2 py-2 pe-4 ps-[3.25rem]'>
-    {Object.keys(colorSuggestions).map(suggestionName => {
-      const color = colorSuggestions[suggestionName as keyof T]?.(baseColor);
-      return (
-        <IconButton
-          key={suggestionName}
-          className={isHexColorLight(color) ? 'text-black' : 'text-white'}
-          style={{ backgroundColor: color }}
-          onClick={() => onSuggestionSelect?.(suggestionName)}
-        >
-          {selectedSuggestion === suggestionName && <DoneIcon />}
-        </IconButton>
-      );
-    })}
-  </div>
 );
 
 const BaseColorsSection = forwardRef<HTMLElement, ComponentProps<'section'>>(
